+++ lib/src/commands/is_upgraded.dart
@@ -11,8 +11,6 @@ import 'package:gg_log/gg_log.dart';
 import 'package:gg_process/gg_process.dart';
 import 'package:gg_status_printer/gg_status_printer.dart';
 import 'package:gg_is_flutter/gg_is_flutter.dart';
-import 'package:matcher/expect.dart';
-import 'package:mocktail/mocktail.dart';
 
 // #############################################################################
 /// Base class for all ggGit commands
@@ -93,31 +91,4 @@ class IsUpgraded extends DirCommand<void> {
 
 // .............................................................................
 /// A Mock for the IsUpgraded class using Mocktail
-class MockIsUpgraded extends Mock implements IsUpgraded {
-  /// Makes [exec] successful or not
-  void mockSuccess({
-    required bool success,
-    required GgLog ggLog,
-    required Directory directory,
-  }) {
-    when(
-      () => exec(
-        directory: any(
-          named: 'directory',
-          that: predicate<Directory>(
-            (d) => d.path == directory.path,
-          ),
-        ),
-        ggLog: any(named: 'ggLog'),
-      ),
-    ).thenAnswer((invocation) async {
-      if (!success) {
-        throw Exception('❌ Upgraded');
-      } else {
-        final ggLog = invocation.namedArguments[const Symbol('ggLog')];
-        ggLog('✅ Upgraded');
-      }
-      return;
-    });
-  }
-}
+class MockIsUpgraded extends MockDirCommand implements IsUpgraded {}